Create search page and commit before using Algolia routes
