+++ app/controllers/search_controller.rb
@@ -1,0 +1,5 @@
+class SearchController < ApplicationController
+  layout "search"
+  def index
+  end
+end
