+++ app/javascript/styles/_kiwi_styles.scss
@@ -15,4 +15,6 @@
 @import './vanilla/organization/text';
 @import './vanilla/organization/font_size';
 @import './vanilla/utilities/button';
+@import './vanilla/utilities/navbar';
+@import './vanilla/utilities/card';
