+++ app/javascript/styles/vanilla/utilities/_card.scss
