+++ app/javascript/styles/vanilla/utilities/_navbar.scss
