+++ app/javascript/styles/vanilla/utilities/_searchbox.scss
@@ -1,0 +1,22 @@
+-webkit-appearance: none;
+background: #f9fafc;
+border: 0;
+border-radius: 6px;
+color: #012d26;
+line-height: normal;
+
+#searchbox {
+  .ais-SearchBox {
+    .ais-SearchBox-form {
+      .ais-SearchBox-input {
+
+      }
+      .ais-SearchBox-submit {
+
+      }
+      .ais-SearchBox-reset {
+        
+      }
+    }
+  }
+}
