+++ app/javascript/views/search/search.scss
@@ -1,5 +1,4 @@
 @import "./../../styles/kiwi_styles";
-@import "./sections/banners";
 
 $fa-font-path: "~@fortawesome/fontawesome-free/webfonts/";
 @import '~@fortawesome/fontawesome-free/scss/fontawesome.scss';
