+++ app/views/layouts/search.html.erb
@@ -1,0 +1,39 @@
+<!DOCTYPE html>
+<html lang="es">
+  <head>
+    <title>Arte Kiwi</title>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="user-scalable=no, initial-scale=1, maximum-scale=1, width=device-width, viewport-fit=cover">
+    <meta name="description" content="Arte Kiwi es una marca de arte, donde transformamos la creatividad en letras y diseños de cajas regalo, anchetas, y mucho más. Somos apasionadas por la letra Timoteo y el lettering.">
+    <meta property="og:url" content="https://artekiwi.com" />
+    <meta property="og:title" content="Arte Kiwi" />
+    <meta property="og:description" content="Arte Kiwi es una marca de arte, donde transformamos la creatividad en letras y diseños de cajas regalo, anchetas, y mucho más. Somos apasionadas por la letra Timoteo y el lettering." />
+
+    <!-- Google Tag Manager -->
+    <!-- <script>(function(w,d,s,l,i){w[l]=w[l]||[];w[l].push({'gtm.start':
+    new Date().getTime(),event:'gtm.js'});var f=d.getElementsByTagName(s)[0],
+    j=d.createElement(s),dl=l!='dataLayer'?'&l='+l:'';j.async=true;j.src=
+    'https://www.googletagmanager.com/gtm.js?id='+i+dl;f.parentNode.insertBefore(j,f);
+    })(window,document,'script','dataLayer','GTM-NQJD44K');</script> -->
+    <!-- End Google Tag Manager -->
+
+    <%= csrf_meta_tags %>
+    <%= csp_meta_tag %>
+    <%= stylesheet_pack_tag 'search_pack' %>
+
+    <!-- Handlee font -->
+    <link href="https://fonts.googleapis.com/css2?family=Handlee&display=swap" rel="stylesheet">
+  </head>
+
+  <body class="font-family-handlee">
+    <!-- Google Tag Manager (noscript) -->
+    <!-- <noscript><iframe src="https://www.googletagmanager.com/ns.html?id=GTM-NQJD44K"
+    height="0" width="0" style="display:none;visibility:hidden"></iframe></noscript> -->
+    <!-- End Google Tag Manager (noscript) -->
+
+    <%= yield %>
+
+    <%#= Include packs with webpacker %>
+    <%= javascript_pack_tag('search_pack', defer: true) %>
+  </body>
+</html>
