+++ test/controllers/search_controller_test.rb
@@ -1,0 +1,9 @@
+require 'test_helper'
+
+class SearchControllerTest < ActionDispatch::IntegrationTest
+  test "should get index" do
+    get search_index_url
+    assert_response :success
+  end
+
+end
